+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - different query start times with multiple queries due to mutating of grafana start time
 
+### Added
+
+- Default Percentile Value when percentile metric is selected
+
 ## 1.1.1
 
 ### Added